requirements.txt
gunicorn
flask
flask_sqlalchemy
psycopg2-binary
flask_login
pandas
python-dotenv
werkzeug